@@ -2,7 +2,7 @@ package com.ysl.baohuo;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.Window;
 import android.view.WindowManager.LayoutParams;
 
